requirements.txt
spacy>=3.0
nltk>=3.8
pymongo>=4.0
openai>=0.27
PyPDF2>=3.0
python-dotenv>=1.0
pytest>=7.0
requests>=2.28
# Optional: if you prefer to use SendGrid SDK
sendgrid>=6.9
